admin panel( класс дедлайна, методы сохранения состояния ввода дедлайна)
