+++ main.py
@@ -6,13 +6,17 @@ from aiogram.enums.parse_mode import ParseMode
 from aiogram.fsm.storage.memory import MemoryStorage
 
 import config
-from handler import router
+import handler_user
+import handler_admin
+import handler_start
 
 
+
+memory = MemoryStorage()
 async def main():
     bot = Bot(token=config.BOT_TOKEN, parse_mode=ParseMode.HTML)
-    dp = Dispatcher(storage=MemoryStorage())
-    dp.include_routers(router)
+    dp = Dispatcher(storage=memory)
+    dp.include_routers(handler_user.router, handler_admin.router, handler_start.router)
     await bot.delete_webhook(drop_pending_updates=True)
     await dp.start_polling(bot, allowed_updates=dp.resolve_used_update_types())
